Rename `test/input-format` -> `test/input-output-format`
