+++ test.js
@@ -7,7 +7,7 @@ require('./test/format-context')
 require('./test/format-context-output')
 require('./test/frame')
 require('./test/image')
-require('./test/input-format')
+require('./test/input-output-format')
 require('./test/packet')
 require('./test/resampler')
 require('./test/stream')
