+++ test/input-output-format.js
